+++ packages/coding-agent/CHANGELOG.md
@@ -8,6 +8,10 @@
 
 - Fixed inherited autocomplete suggestions to refresh after editor cursor movement ([#5499](https://github.com/earendil-works/pi/pull/5499) by [@Roman-Galeev](https://github.com/Roman-Galeev)).
 
+### Removed
+
+- Removed todo continuation auto-follow-up (the `--disable-todo-continuation` flag and `todotools.continuation` settings).
+
 ## [2026.6.6-3] - 2026-06-06
 
 ### Added